requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pydantic==2.5.2
sqlite3
python-dateutil==2.8.2
schedule==1.2.1
